requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.15
python-dotenv>=1.0.0
openai>=1.0.0
typing-extensions>=4.8.0
pydantic>=2.0.0 